Fix a warning for certain compilers.
NOTE: ssize_t can be 32 or 64 bit (depending on platform), so just casting
that to a int in this case since we aren't going to read in > 2^32-1 bytes.
close ticket:3080

+++ lib/netplay/netplay.cpp
@@ -2221,7 +2221,7 @@ static void NETallowJoining(void)
 		}
 		else
 		{
-			debug(LOG_NET, "Failed to process joining, socket not ready or no data, recv_result is :%ld", recv_result);
+			debug(LOG_NET, "Failed to process joining, socket not ready or no data, recv_result is :%d", (int)recv_result);
 			connectFailed = true;
 		}
